feat(decomp): advance Masu engine core
Recover twelve additional strict-exact Masu functions and the target data/sdata2 ownership without padding or register forcing.

Object gate: objdiff strict 115/119 exact; functionRelocDiffs=data_value 115/119 exact; text 0x5DC8/0x5DB0 at 99.614460%; data 0x128 and sdata2 0x90 exact.

No Matching flip: MasuDraw, MasuFind, mbMasuPlayerColorSet, and mbMasuPKinokoValueGet remain divergent, so the fallback is retained and no full-DOL gate is claimed.

+++ include/game/board/masu.h
@@ -59,7 +59,7 @@ void mbMasuTypeSet(s16 id, int type);
 void mbMasuTypeChange(u16 oldType, u16 newType);
 BOOL mbMasuDispCheck(s16 id);
 void mbMasuCapsuleReset(void);
-s16 mbMasuCapsuleGet(s16 id);
+int mbMasuCapsuleGet(s16 id);
 void mbMasuCapsuleSet(s16 id, int capsuleNo);
 void mbMasuPosGet(s16 id, HuVecF *pos);
 void mbMasuPosSet(s16 id, float x, float y, float z);
